# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Phishing Simulation Platform API",
        "description": "Comprehensive API for managing phishing simulation campaigns, employee training, and security awareness programs.",
        "contact": {
            "email": "support@phishingsim.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.2:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/company/dashboard": {
            "get": {
                "tags": [
                    "Company"
                ],
                "summary": "Get company dashboard",
                "description": "Get comprehensive dashboard data including company information, statistics, and recent activity",
                "operationId": "43f0c2e4c2d1377b58afaeaf4a3eb71d",
                "responses": {
                    "200": {
                        "description": "Dashboard data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dashboard data retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "company": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Acme Corporation"
                                                        },
                                                        "email": {
                                                            "type": "string",
                                                            "example": "admin@acme.com"
                                                        },
                                                        "plan_id": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "plan": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer",
                                                                    "example": 2
                                                                },
                                                                "name": {
                                                                    "type": "string",
                                                                    "example": "Basic"
                                                                },
                                                                "price": {
                                                                    "type": "number",
                                                                    "format": "float",
                                                                    "example": 10
                                                                },
                                                                "employee_limit": {
                                                                    "type": "integer",
                                                                    "example": 50
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "statistics": {
                                                    "properties": {
                                                        "total_campaigns": {
                                                            "type": "integer",
                                                            "example": 5
                                                        },
                                                        "active_campaigns": {
                                                            "type": "integer",
                                                            "example": 2
                                                        },
                                                        "completed_campaigns": {
                                                            "type": "integer",
                                                            "example": 3
                                                        },
                                                        "total_targets": {
                                                            "type": "integer",
                                                            "example": 150
                                                        },
                                                        "total_interactions": {
                                                            "type": "integer",
                                                            "example": 450
                                                        },
                                                        "successful_simulations": {
                                                            "type": "integer",
                                                            "example": 4
                                                        },
                                                        "vulnerable_employees": {
                                                            "type": "integer",
                                                            "example": 12
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "recent_activity": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "type": {
                                                                "type": "string",
                                                                "example": "campaign_completed"
                                                            },
                                                            "message": {
                                                                "type": "string",
                                                                "example": "Phishing Campaign #3 completed"
                                                            },
                                                            "timestamp": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Error message"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "field": [
                                "Validation error message"
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT token in format (Bearer <token>)"
            }
        }
    },
...
}